+++ src/array.rs
@@ -67,7 +67,7 @@ use crate::{
 ///
 /// ```
 /// # rosy::vm::init().unwrap();
-/// # use rosy::Array;
+/// # use rosy::prelude::*;
 /// let array: Array = (0..10).collect();
 ///
 /// assert_eq!(array.to_s(), "[0, 1, 2, 3, 4, 5, 6, 7, 8, 9]");